requirements.txt
# Backend - FastAPI
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv

# Dashboard - Plotly Dash
dash
plotly
pandas

# Testing
pytest
httpx
faker
